Removed check for empty file name from property editor/setter.  Now an exception is thrown in the doit() method but not when the property dialog is closed.
The "correct" thing would be to insist the user enter a filename but since
the module info is tied to the module property editor, it's really annoying
if the user just wants to see the info but not yet enter a filename.  also,
when buiding nested itineraries you often want to have these modules but not
enter the filename.

So, at least for now, the check isn't done until doit().   Left code in module
but commented it out in case later there's a way to do the check correctly w/o
being annoying.